1. 优化了item显示结构 2. 增加了一个预览pdf的小功能（）
