+++ package.json
@@ -11,12 +11,13 @@
   },
   "dependencies": {
     "@fontsource-variable/saira": "^5.0.29",
-    "@radix-ui/react-dialog": "^1.0.5",
+    "@radix-ui/react-dialog": "^1.1.1",
     "@radix-ui/react-icons": "^1.3.0",
     "@radix-ui/react-slot": "^1.0.2",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
     "fuse.js": "^7.0.0",
+    "jszip": "^3.10.1",
     "next-themes": "^0.3.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
